stable V with the correct specular

+++ sources/lighting/phong.c
@@ -6,7 +6,7 @@
 /*   By: azgaoua <azgaoua@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/29 18:19:01 by hlaadiou          #+#    #+#             */
-/*   Updated: 2024/03/01 16:39:16 by hlaadiou         ###   ########.fr       */
+/*   Updated: 2024/03/02 15:47:19 by azgaoua          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,7 +56,7 @@ t_color	illuminate(t_object *obj, t_point px, t_light light, t_point cam)
 	float		reflect_dot_eye;
 	float		factor;
 
-	px_color = schur_product(_color(1.0, 0.0, 0.0), light.color);
+	px_color = schur_product(_color((float)1, (float)0, (float)0), light.color);
 	ph = _phong(obj, px, light, cam);
 	ambient = multiply_color_scalar(obj->specs.ambient, px_color);
 	light_dot_normal = dot_product(ph.l, ph.n);
